apps/api/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "api" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "tests.py"]
addopts = """
 --cov
 --cov-report html:'../coverage/apps/html'
 --cov-report xml:'../coverage/apps/coverage.xml'
 --html='../reports/apps/unittests/html/index.html'
 --junitxml='../reports/apps/unittests/junit.xml'
"""

[tool.poetry]
name = "api"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "api"

  [tool.poetry.dependencies]
  python = ">=3.12,<3.15"
  django = "^6.0"
  django-cors-headers = "^4.4"
  pillow = "^11.0"  # For image handling
  python-decouple = "^3.8"  # For environment variables
  djangorestframework = "^3.17.1"
  drf-spectacular = "^0.29.0"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.3.1"
  pytest = "8.3.4"
  pytest-django = "4.9.0"
  pytest-sugar = "1.0.0"
  pytest-cov = "6.0.0"
  pytest-html = "4.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
